third_party/blink/web_tests/fast/js/kde/garbage-n-expected.txt

CONSOLE ERROR: Uncaught SyntaxError: Unexpected identifier 'code'
KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Got expected error: 'SyntaxError'
PASS successfullyParsed is true

TEST COMPLETE
